refactor(accounts): modularize postgres transaction posting tests
- Centralize ledger mock, DB fixtures, and claim helpers under tests/postgres_transaction_posting/support for lower duplication (Sonar).
- Split scenarios into focused modules (try_claim, claim_batch, process_claimed, find_by_id, update_status); remove tests/common in favor of the integration test directory layout.
